requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
jinja2==3.1.2
aiofiles==23.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
slowapi==0.1.9
redis==5.0.1
requests==2.31.0

